third_party/WebKit/LayoutTests/http/tests/inspector/elements/styles/update-locations-on-filesystem-scss-load.html

<html>
<head>
<script src="../../inspector-test.js"></script>
<script src="../../isolated-filesystem-test.js"></script>
<script src="../../workspace-test.js"></script>
<script>
function loadCSS()
{
    var link = document.createElement("link");
    link.rel = "stylesheet";
    link.type = "text/css";
    link.href = "resources/update-locations-on-filesystem-scss-load.css";
    document.head.appendChild(link);
}

function test()
{
    InspectorTest.addResult("Creating filesystem with the SCSS file...");
    var fs = new InspectorTest.TestFileSystem("/var/www");
    fs.root.addFile("update-locations-on-filesystem-scss-load.scss", ["a {", "  foo: bar;", "/* COMMENT */", "  font-size: 12px;", "}"].join("\n"));
    fs.addFileMapping(WebInspector.ParsedURL.completeURL(InspectorTest.resourceTreeModel.inspectedPageURL(), "resources/source/"), "/");
    fs.reportCreated(fileSystemCreated);

    function fileSystemCreated()
    {
        InspectorTest.addResult("Loading raw css with mapping...");
        InspectorTest.cssModel.addEventListener(WebInspector.CSSStyleModel.Events.StyleSheetAdded, styleSheetAdded);
        InspectorTest.evaluateInPage("loadCSS()");
    }

    var liveLocation;

    function styleSheetAdded(event)
    {
        InspectorTest.addResult("Stylesheet was added, dumping location:");
        var header = event.data;
        var cssLocation = new WebInspector.CSSLocation(header.cssModel(), header.id, header.sourceURL, 0, 1);
        liveLocation = WebInspector.cssWorkspaceBinding.createLiveLocation(cssLocation, function() {});
        InspectorTest.addSniffer(WebInspector.CSSWorkspaceBinding.prototype, "pushSourceMapping", afterBind);
        dumpLiveLocation();
    }

    function afterBind()
    {
        InspectorTest.addResult("Source map was bound to the file loaded from filesystem:");
        dumpLiveLocation();
        InspectorTest.completeTest();
    }

    function dumpLiveLocation()
    {
        var uiLocation = liveLocation.uiLocation();
        if (!uiLocation) {
            InspectorTest.addResult("Null uiLocation");
            return;
        }
        var networkURL = WebInspector.networkMapping.networkURL(uiLocation.uiSourceCode);
        InspectorTest.addResult(networkURL + ":" + uiLocation.lineNumber + ":" + uiLocation.columnNumber);
    }
};
</script>
</head>
<body onload="runTest()">
<p>Tests that stylesheet LiveLocations are updated when an SCSS file is loaded from the filesystem.</p>
</body>
</html>
