third_party/WebKit/LayoutTests/http/tests/inspector/elements/styles/update-locations-on-filesystem-scss-load-expected.txt

Tests that stylesheet LiveLocations are updated when an SCSS file is added from the filesystem.

Adding file system.
Adding stylesheet header.
Adding network resource.
LiveLocation updated:
http://127.0.0.1:8000/inspector/elements/styles/resources/update-locations-on-filesystem-scss-load.css:0:1
Adding filesystem SCSS.
LiveLocation updated:
http://127.0.0.1:8000/inspector/elements/styles/resources/update-locations-on-filesystem-scss-load.scss:0:2
